refactor: ServerSSLHandler
