+++ zio-http/src/main/scala/zhttp/service/server/OptionalServerSSLHandler.scala
@@ -8,7 +8,7 @@ import io.netty.handler.ssl.{SslContext, SslHandler}
 import zhttp.service.Server.Config
 import zhttp.service._
 import zhttp.service.server.ServerChannelInitializerUtil.configureClearTextHttp2
-import zhttp.service.server.ServerSSLHandler.SSLHttpBehaviour
+import zhttp.service.server.ServerSSLBuilder.SSLHttpBehaviour
 
 import java.util
