+++ zio-http/src/test/scala/zhttp/service/SSLSpec.scala
@@ -4,7 +4,7 @@ import io.netty.handler.codec.DecoderException
 import io.netty.handler.ssl.SslContextBuilder
 import zhttp.http._
 import zhttp.service.client.ClientSSLHandler.ClientSSLOptions
-import zhttp.service.server.ServerSSLHandler.{ServerSSLOptions, ctxFromCert}
+import zhttp.service.server.ServerSSLBuilder.{ServerSSLOptions, ctxFromCert}
 import zhttp.service.server._
 import zio.ZIO
 import zio.duration.durationInt
